requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
beautifulsoup4==4.11.1
lxml==4.9.1
requests==2.28.1
psycopg2-binary==2.9.4
requests==2.28.1
Django==4.1.2
djangorestframework==3.14.0
psycopg==3.1.3
psycopg2-binary==2.9.4
sqlalchemy